ai-customer-service/backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
aiofiles